Update extract_feature.py
The feature set used to build the 'concat' predictor is [ HW, CIN1 , CIN2 , CIN3 , CIN4 ] but when we are predicting , the feature set is wrongly extracted for concat layer. The ' no. of input tensors ' feature value is not required in the feature list because we are adding padding values to the list if no. of tensors are less than 4 and removing extra tensors if no. of tensors more than 4 thus maintaining the consistency. Even no. of channels are wrongly extracted thus made changes accordingly.